+++ agents/mongodb/src/version/v3_2/sharded.rs
@@ -74,10 +74,6 @@ impl Agent for Sharded {
         }
     }
 
-    fn service_name(&self) -> String {
-        "mongod".into()
-    }
-
     fn shards(&self, span: &mut Span) -> Result<Shards> {
         if self.is_mongos {
             Ok(Shards::new(Vec::new()))